requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1 ; python_version >= '3.8'
django==3.2.20
django-mysql==4.11.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1 ; python_version >= '3.5'
mysqlclient==2.2.0
packaging==24.1 ; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1 ; python_version >= '3.6'
sqlparse==0.5.0 ; python_version >= '3.8'
uritemplate==4.1.1 ; python_version >= '3.6'
